# swagger.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "CTinyURL API",
    "description": "demo tiny url.",
    "version": "0.0.1",
    "contact": {}
  },
  "tags": [
    {
      "name": "tinyurl",
      "description": ""
    }
  ],
  "servers": [],
  "components": {
    "securitySchemes": {
      "bearer": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http"
      }
    },
    "schemas": {
      "GetShortLinkReponseDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "url": {
            "type": "string"
          },
          "clicksCount": {
            "type": "number"
          }
        },
        "required": [
          "id",
          "url",
          "clicksCount"
        ]
      },
      "GetShortLinkDetailsReponseDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "url": {
            "type": "string"
          },
          "clicksCount": {
            "type": "number"
          },
          "clicks": {
            "type": "array",
            "items": {
              "type": "string"
            }
          }
        },
        "required": [
          "id",
          "url",
          "clicksCount",
          "clicks"
        ]
      },
      "CreateLinkRequestDto": {
        "type": "object",
        "properties": {
          "url": {
            "type": "string"
          }
        },
        "required": [
          "url"
        ]
      }
    }
  },
  "paths": {
    "/ping": {
      "get": {
        "operationId": "UrlController_getHello",
        "summary": "",
        "description": "ping",
        "parameters": [],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "Root"
        ]
      }
    },
    "/{id}": {
      "get": {
        "operationId": "redirectTo",
        "summary": "",
        "description": "redirect to target url, if exist",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "301": {
            "description": "Redirect to target, if exist"
          },
          "302": {
            "description": "Redirect to default"
          }
        },
        "tags": [
          "Root"
        ]
      }
    },
    "/link/{id}": {
      "get": {
        "operationId": "getShortLink",
        "summary": "",
        "description": "return basic url object, if exist",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "basic url object",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GetShortLinkReponseDto"
                }
              }
            }
          },
          "404": {
            "description": "Not found"
          }
        },
        "tags": [
          "link"
        ]
      }
    },
    "/link/{id}/details": {
      "get": {
        "operationId": "getShortLinkDetails",
        "summary": "",
        "description": "return details url object, with details of every click, if exist",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "details url object",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GetShortLinkDetailsReponseDto"
                }
              }
            }
          },
          "404": {
            "description": "Not found"
          }
        },
        "tags": [
          "link"
        ]
      }
    },
    "/link": {
      "post": {
        "operationId": "createShortLink",
        "summary": "",
        "description": "create url short link",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateLinkRequestDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "success create object",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CreateLinkRequestDto"
                }
              }
            }
          }
        },
        "tags": [
          "link"
        ]
      }
    }
  }
}